Update Shahar/Philip system design
Devops is an overloaded term that includes cultural apects that wouldn't
necessarily be discussed in a system design interview.

Hashisuite is a particular toolset. While we have experience with it
there are not other tools sets listed for design so we can keep this
focused on what it does, not what it is.